refactor: strengthen type safety — remove unnecessary static_casts, use StrongIndexVector and as_label
- Replace 16x std::to_string(static_cast<int>(...)) with as_label() in element_column_resolver
- Make in_range() template accept only strong types (is_strong_type constraint)
- Fix StrongIndexVector operator[]/at() return types to support vector<bool>
- Type PhaseGridRecorder::record() with IterationIndex, SceneUid, PhaseIndex
- Convert sddp_scene_tracker vectors to StrongIndexVector<SceneIndex, ...>
- Replace 6 raw loops with std::ranges algorithms in sddp_scene_tracker
- Migrate enumerate<SceneIndex>/enumerate<PhaseIndex> in planning_lp, linear_interface
- Remove ~30 unnecessary static_cast<int> across SDDP forward/backward/aperture passes

+++ source/monolithic_method.cpp
@@ -102,8 +102,8 @@ auto MonolithicMethod::solve(PlanningLP& planning_lp, const SolverOptions& opts)
           (std::filesystem::path(lp_debug_directory) / "gtopt_lp").string();
       for (const auto& phase_systems : planning_lp.systems()) {
         for (const auto& system : phase_systems) {
-          const auto su = static_cast<int>(system.scene().uid());
-          const auto pu = static_cast<int>(system.phase().uid());
+          const auto su = system.scene().uid();
+          const auto pu = system.phase().uid();
           if (!in_lp_debug_range(su,
                                  pu,
                                  lp_debug_scene_min,
